# Introduce conditional merge-operator invocation in point lookups